+++ articlear/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter } from 'react-router-dom'
 import './App.css'
 import RouterConfig from './config/RouterConfig'
 import TopBar from './components/Topbar'
-import Footer1 from './components/Footer/footer'
+import Footer1 from './components/footer'
 
 const theme = createTheme({
     type: 'dark', // it could be "light" or "dark"
@@ -21,7 +21,7 @@ const theme = createTheme({
             blue9: '#68ddfd',
             blue10: '#8ae8ff',
             textLowContrast: '#2ec8ee',
-            textHighContrast: '#eaf8ff'
+            textHighContrast: '#eaf8ff',
         },
         space: {},
         fonts: {},
@@ -35,7 +35,7 @@ function App() {
                 <BrowserRouter>
                     <TopBar />
                     <RouterConfig />
-                    <Footer1 />
+                    <Footer1 className="sticky bottom-0" />
                 </BrowserRouter>
             </NextUIProvider>
         </>
